python/requirements.txt
pandas
psycopg2-binary
openpyxl
tqdm
python-dotenv